Complete sorting feature
